# Specifications/openapi/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Student Residence Service",
    "description" : "An API that allows management of rentals of home appliance and management of announcements",
    "version" : "1.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:9852"
  } ],
...
  "paths" : {
...
    "/v1/appliances/{appliance-id}" : {
      "get" : {
        "tags" : [ "Home Appliance" ],
        "description" : "Returns a specific house appliance",
        "operationId" : "getAppliance",
        "parameters" : [ {
          "name" : "appliance-id",
          "in" : "path",
          "description" : "id of the house appliance",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "example" : 1
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Specific house appliance information",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Appliance"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request. Please check the parameters"
          },
          "401" : {
            "description" : "Unauthorized. Please check your credentials"
          },
          "403" : {
            "description" : "Forbidden. Sorry, you can not access this resource."
          }
        }
      },
      "put" : {
        "tags" : [ "Home Appliance" ],
        "description" : "Updates a specific House appliance",
        "parameters" : [ {
          "name" : "appliance-id",
          "in" : "path",
          "description" : "id of the house appliance",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "example" : 1
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Appliance"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successfully updated the house appliance"
          },
          "401" : {
            "description" : "Unauthorized. Please check your credentials"
          },
          "403" : {
            "description" : "Forbidden. Sorry, you can not access this resource."
          }
        }
      },
      "delete" : {
        "tags" : [ "Home Appliance" ],
        "description" : "Updates a specific House appliance",
        "parameters" : [ {
          "name" : "appliance-id",
          "in" : "path",
          "description" : "id of the house appliance",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "example" : 1
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Successfully deletion of the house appliances"
          },
          "401" : {
            "description" : "Unauthorized. Please check your credentials"
          },
          "403" : {
            "description" : "Forbidden. Sorry, you can not access this resource."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Appliance_type" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
      "Appliance" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "serial_number" : {
            "type" : "integer"
          },
          "model_name" : {
            "type" : "string"
          },
          "Appliance_type" : {
            "$ref" : "#/components/schemas/Appliance_type"
          },
          "state" : {
            "type" : "string",
            "format" : "enum - AVAILABLE - BROKEN"
          },
          "price_per_day" : {
            "type" : "number",
            "format" : "float"
          },
          "max_time" : {
            "type" : "integer"
          },
          "available_appliances" : {
            "type" : "integer"
          },
          "creation_date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deleted_on" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    },
...
    "securitySchemes" : {
      "oAuth2Password" : {
        "type" : "oauth2",
        "description" : "OAuth 2.0 Security Scheme",
        "flows" : {
          "password" : {
            "tokenUrl" : "/oauth/token",
            "scopes" : { }
          }
        }
      }
    },
...
  }
}